Move the now shared author finding code out
This code is used in the new external links code and in the existing
jinja extension. It is now moved into a shared taxonomy module which
makes a bit more sense.
