+++ ckanext/nhm/lib/taxonomy.py
@@ -1,8 +1,7 @@
 from collections import OrderedDict
 
+import abc
 import re
-
-from ckanext.nhm.lib.jinja_extensions import AuthorParserStage, SimpleFieldParserStage, CapitalisedParserStage
 
 
 def extract_ranks(record):
@@ -78,28 +77,152 @@ def extract_species(record):
     :param record: the record dict
     :return: the species value, or None if it is not present
     '''
-    # first try extracting the species from the scientific name, which starts with the species but often has an author
-    # or date after it
+    # try extracting the species from the scientific name, which starts with the species but often has an author or date
+    # after it
     scientific_name = record.get('scientificName', None)
     if scientific_name:
-        # if it's just a single word, no need to parse
-        first_space = re.search('\s', scientific_name)
-        if not first_space:
-            return scientific_name
-
-        evaluators = OrderedDict()
-        evaluators['authors'] = AuthorParserStage()
-        evaluators['species'] = SimpleFieldParserStage('specificEpithet')
-        evaluators['subgenus'] = SimpleFieldParserStage('subgenus')
-        evaluators['capitalisation'] = CapitalisedParserStage()
-
-        ix = None
-        for t, e in evaluators.items():
-            ix = e.evaluate(scientific_name, record)
-            if ix:
-                break
-
-        if ix:
-            return scientific_name[:ix].strip()
+        ix = find_author_split(scientific_name, record)
+        return scientific_name[:ix].strip()
 
     return None
+
+
+def find_author_split(value, record_dict):
+    '''
+    Given a string and a record attempts to determine where in the string an author is defined. If an author is found
+    then this will return the index the author part starts, otherwise returns None.
+    :param value:           the value to search in
+    :param record_dict:     the record dictionary to use as a supplementary source of information
+    :return: the index at the start of the author part or None
+    '''
+    first_space = re.search('\s', value)
+    if not first_space:
+        return None
+
+    evaluators = [
+        AuthorParserStage(),
+        SimpleFieldParserStage('specificEpithet'),
+        SimpleFieldParserStage('subgenus'),
+        CapitalisedParserStage()
+    ]
+
+    ix = None
+    for evaluator in evaluators:
+        ix = evaluator.evaluate(value, record_dict)
+        if ix:
+            break
+
+    return ix
+
+
+class BaseParserStage(object):
+    '''
+    Represents a single stage of a field parsing process.
+    '''
+
+    @abc.abstractmethod
+    def _meets_criteria(self, body, record_dict):
+        '''
+        Tests to see if the item meets certain criteria.
+        :param body: the tag body to search in
+        :param record_dict: the full record
+        :return: boolean for pass/fail
+        '''
+        return True
+
+    @abc.abstractmethod
+    def _extract(self, body, record_dict):
+        '''
+        Finds an index within a string
+        :param body: the tag body
+        :param record_dict: the full record
+        :return: a character index
+        '''
+        return 0
+
+    def evaluate(self, body, record_dict):
+        '''
+        Checks if the item meets the criteria then returns the index
+        :param body: the tag body
+        :param record_dict: the full record
+        :return: character index if criteria met, None if not
+        '''
+        if self._meets_criteria(body, record_dict):
+            return self._extract(body, record_dict)
+
+
+class AuthorParserStage(BaseParserStage):
+    '''
+    A specific stage for searching for authors in the tag body.
+    '''
+
+    def _meets_criteria(self, body, record_dict):
+        '''
+        Ensures an author field is present in the record.
+        '''
+        return 'scientificNameAuthorship' in record_dict.keys()
+
+    def _extract(self, body, record_dict):
+        '''
+        Searches for the full author string, then breaks it up into smaller pieces (sections in brackets, individual
+        names) if that's not found
+        :return: the start index of the author string if found, otherwise None
+        '''
+        full_author = record_dict['scientificNameAuthorship']
+        author_strings = [full_author] + [p.strip() for p in set(
+            re.findall(u'\(([\w\s]+)\)', full_author) + re.findall(
+                u'([\w.\s]+)', full_author))]
+        for a in author_strings:
+            matches = re.search(u'\s\(?{0}\)?(\s|$)'.format(re.escape(a)), body)
+            return matches.start() if matches else None
+
+
+class SimpleFieldParserStage(BaseParserStage):
+    '''
+    A generic stage for searching for a certain field within the tag body (for the purposes of finding authors).
+    '''
+
+    def __init__(self, field_name):
+        self.field_name = field_name
+
+    def _meets_criteria(self, body, record_dict):
+        '''
+        Checks that the record contains a value for this field and that the value is present in the tag body.
+        '''
+        return self.field_name in record_dict.keys() and record_dict[
+            self.field_name] in body
+
+    def _extract(self, body, record_dict):
+        '''
+        If the value is at the end of the string, there is no point in continuing; otherwise, it looks for the first
+        capitalised word after that value.
+        :return: the start index of the estimated author string if found, else None.
+        '''
+        field_value = record_dict[self.field_name]
+        if re.search('{0}$'.format(re.escape(field_value)), body):
+            return len(body)
+        split_by_value = re.split('{0}'.format(re.escape(field_value)), body,
+                                  1)
+        matches = re.search('\(?[A-Z]\w*', split_by_value[1])
+        return matches.start() + len(split_by_value[0]) + len(
+            field_value) if matches else None
+
+
+class CapitalisedParserStage(BaseParserStage):
+    '''
+    The last resort stage in the search for authors - searches for the second capitalised word in the tag body.
+    '''
+
+    def _meets_criteria(self, body, record_dict):
+        '''
+        Checks for multiple capitalised words in the tag body.
+        '''
+        capit = re.findall('([A-Z]\S*)(?:\s|$)', body)
+        return len(capit) > 1
+
+    def _extract(self, body, record_dict):
+        '''
+        Finds the start index of the second capitalised word.
+        '''
+        matches = [m for m in re.finditer('[A-Z]', body)]
+        return matches[1].start()
